skyrim_rag_project/requirements.txt
pandas
numpy
transformers
torch
faiss-cpu